add CategoryHeadings to result page
