+++ public/index.html
@@ -24,7 +24,7 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <link
       rel="stylesheet"
-      href="https://fonts.googleapis.com/css2?family=Open+Sans:wght@300;400;700&family=Roboto:wght@300;400;500;700&family=Roboto+Slab:wght@700&display=swap"
+      href="https://fonts.googleapis.com/css2?family=Open+Sans:wght@300;400;700&family=Roboto:wght@300;400;500;700&family=Roboto+Slab:wght@400;700&display=swap"
     />
     <title>MyFriendBen</title>
   </head>
