+++ src/Components/Results/CategoryHeading/CategoryHeading.css
@@ -18,7 +18,6 @@
 .text-style {
   font-family: 'Roboto Slab', serif;
   font-weight: 700;
-  font-size: 1.25rem;
   color: var(--secondary-color);
 }
